refactor: extract is_test_path primitive, sever the extension->code_audit cycle (#8425)
Keystone for a future homeboy-audit extraction. code_audit is otherwise a clean
top-consumer (component->audit: 0, engine->audit: 0), blocked only by a single
extension->code_audit cycle edge. That edge was three references, all to things
that are actually low-level primitives, not audit domain logic:

- extension/test/drift.rs imported code_audit::walker::is_test_path (a pure,
  zero-dependency test-path classifier) — twice (a use + an inline call).
- extension/manifest.rs called code_audit::conventions::Language::builtin_* —
  but Language already lives in homeboy-engine-primitives (this campaign);
  code_audit::conventions just re-exports it.

Move is_test_path into homeboy-engine-primitives as `test_path` (pure string
classification, no deps), re-export it from code_audit::walker + code_audit root
so the ~18 existing call sites are unchanged, and repoint extension's three refs
straight at the primitives crate.

Result: `extension -> code_audit` real edges drop to ZERO. code_audit now sits
cleanly above component/extension/engine with no back-edge — the last structural
blocker to extracting it as homeboy-audit.

Verified: primitives + core compile clean; 3 test_path tests pass;
`cargo check --workspace --tests` green.

Refs #8425
