+++ crates/homeboy-core/src/code_audit/mod.rs
@@ -75,9 +75,9 @@ pub(crate) use execution_plan::{
 };
 pub use findings::{homeboy_finding_from_audit, Finding, FindingConfidence, Severity};
 pub use fingerprint::FileFingerprint;
+pub use homeboy_engine_primitives::test_path::is_test_path;
 pub use report::AuditCommandOutput;
 pub use run::{run_main_audit_workflow, AuditRunWorkflowArgs, AuditRunWorkflowResult};
-pub use walker::is_test_path;
 
 pub use entry::{
     audit_component, audit_path, audit_path_scoped, audit_path_with_id,
